+++ nvim/after/plugin/lsp.lua
@@ -1,4 +1,6 @@
 local cmp = require("cmp")
+local Remap = require("jrbb.keymap")
+local nnoremap = Remap.nnoremap
 
 -- Disable inline messages
 vim.diagnostic.config({virtual_text = false})
@@ -37,3 +39,18 @@ require("lspconfig").rust_analyzer.setup({
 -- npm i -g typescript typescript-language-server
 require("lspconfig").tsserver.setup(config());
 
+nnoremap("<leader>gd", function() vim.lsp.buf.definition() end)
+nnoremap("<leader>d", function() vim.lsp.buf.hover() end)
+nnoremap("<leader>gi", function() vim.lsp.buf.implementation() end)
+nnoremap("<leader>gs", function() vim.lsp.buf.signature_help() end)
+nnoremap("<leader>gy", function() vim.lsp.buf.type_definition() end)
+nnoremap("<leader>gr", function() vim.lsp.buf.references() end)
+nnoremap("<leader>rn", function() vim.lsp.buf.rename() end)
+nnoremap("<leader>gg", function()
+  vim.diagnostic.setqflist()
+  vim.cmd("cclose")
+  vim.cmd("cnext")
+end)
+nnoremap("<leader>gl", function() vim.diagnostic.setloclist() end)
+nnoremap("<C-k>", function() vim.diagnostic.goto_next() end)
+nnoremap("<C-j>", function() vim.diagnostic.goto_prev() end)
