+++ nvim/lua/jrbb/plugins.lua
